refactor: split storage engine into focused submodules
Split src/storage/engine.rs (3,817 lines) into 8 smaller, focused modules:

- mod.rs (142 lines): Core StorageEngine struct and public API
- error.rs (43 lines): Error enum and From implementations
- read.rs (192 lines): get_account, get_storage, get_value_from_page
- write.rs (282 lines): set_values, set_values_in_page, set_values_in_cloned_page
- handlers.rs (709 lines): Trie modification handlers (branch/leaf operations)
- page_ops.rs (281 lines): allocate, orphan, clone, split pages
- helpers.rs (161 lines): Utility functions (prefix matching, subtrie ops)
- tests.rs (2,286 lines): Extracted test module
